Added div's for new Home page. Modified old Home files to comment out code for reference usage.
