+++ frontend/src/Pages/Home.tsx
@@ -1,33 +1,15 @@
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import CustomButton from '../Components/CustomButton';
-  
-  const Home = () => {
-    const [count, setCount] = useState(1);
-  
-    const handleButtonPress = () => {
-      setCount(count + 1);
-    }
-  
-    const nextPage = useNavigate();
-  
-    return (
-      <div className="App">
-        <header className="App-header"> {/* Header */}
-          <h1>React.js Header</h1> 
-        </header>
-        <body className="App-body"> {/* Body */}
-          <p>
-            Hey. <br/> This is just a body test. <br/> Cool, huh? :)
-          </p>
-        </body>
-        <div className="App-footer"> {/* Footer */}
-          <div className='footer-item'><CustomButton title="Do Not Press" disabled={false} click={handleButtonPress}/></div>
-          <div>{count}</div>
-          <div className='footer-item'><CustomButton title="Next Page" disabled={false} click={() => nextPage('TestPage')}/></div>
-        </div>
-      </div>
-    );
-  }
+import { useNavigate } from "react-router-dom";
 
-  export default Home;
+
+const Home = () => {
+    return (
+        <div className="Homepage">
+            <div className="Homepage-header"></div>
+            <div className="Homepage-columnLeft"></div>
+            <div className="Homepage-columnRight"></div>
+            <div className="Homepage-footer"></div>
+        </div>
+    );
+}
+
+export default Home;
